requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
attrs==23.2.0
black==24.4.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.6.2
charset-normalizer==3.3.2
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==8.1.7
cloudinary==1.40.0
cloudpathlib==0.18.1
confection==0.1.5
cymem==2.0.8
distro==1.9.0
dj-database-url==2.2.0
Django==4.2.11
django-cloudinary-storage==0.3.0
django-jsonfield==1.4.1
django-ranged-response==0.2.0
django-secure-contact-form==0.2
django-simple-captcha==0.6.0
django-stripe-payments==2.0.0
djangorestframework==3.15.1
#en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
flake8==7.1.0
flexcache==0.3
flexparser==0.3.1
frozenlist==1.4.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mathparse==0.1.2
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.4
packaging==24.1
pathspec==0.12.1
pillow==10.3.0
Pint==0.24.3
platformdirs==4.2.2
pluggy==1.5.0
preshed==3.0.9
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
Pygments==2.18.0
pymongo==3.13.0
pytest==8.2.2
pytest-django==4.8.0
python-dateutil==2.7.5
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==3.13
regex==2024.5.15
requests==2.31.0
rich==13.7.1
sentry-sdk==2.5.1
setuptools==71.0.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.2.19
sqlparse==0.5.0
srsly==2.4.8
stripe==9.6.0
thinc==8.2.5
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.11.0
urllib3==2.2.1
wasabi==1.1.3
weasel==0.4.1
whitenoise==6.6.0
wrapt==1.16.0
yarl==1.9.4
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889